+++ SplitWiser/src/main/java/com/splitwiser/SplitWiser/user/UserRepository.java
@@ -1,9 +1,10 @@
 package com.splitwiser.SplitWiser.user;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     // define queries here
 }
